backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.3
charset-normalizer==3.3.2
click==8.1.7
dnspython>=1.16.0
email-validator==2.1.1
exceptiongroup==1.2.0
fast-api-gcs==0.0.11
fastapi==0.110.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
motor==3.4.0
orjson==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pymongo==4.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
